refactor project_config client
