+++ microservices/requestApi/routes/v1/clients/project_config_client.js
@@ -1,1 +1,1 @@
-module.exports = require('../../../../projectApi/clients/js/project_config_client');
+module.exports = require('../../../../shared/js/project_config_client');
