+++ microservices/shared/js/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "ocwa-clients",
+  "author": "Jeremy Ho",
+  "version": "1.0.0",
+  "license": "Apache-2.0",
+  "private": true,
+  "scripts": {
+    "clean": "rm -rf coverage dist",
+    "purge": "rm -rf node_modules",
+    "start": "node ./bin/www",
+    "test": "NODE_ENV=test nyc --reporter=lcov mocha --exit --timeout 10000"
+  },
+  "dependencies": {
+    "config": "^3.0.1",
+    "morgan": "^1.9.1",
+    "npmlog": "^4.1.2",
+    "request": "^2.88.0"
+  },
+  "devDependencies": {
+    "chai": "^4.1.2",
+    "chai-http": "^4.2.0",
+    "mocha": "^6.1.2",
+    "nyc": "^13.3.0",
+    "sinon": "^7.3.1"
+  }
+}
