+++ microservices/shared/js/src/project_config_client.js
